got a technical version of the shelf maker ui working.
It was really really dumb, I kept messing up sizeDelta for the rect transform.

Either way, I am probably going to rebuild this again. But make it so you have to give coords to how big each shelf is because right as finishing this, I realized this does not give enough control.